backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx>=0.24.0
pydantic-settings==2.1.0
email-validator==2.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
mutagen==1.47.0
supabase>=2.0.0
openai==1.12.0

